refactor: change value Int constructor to Int64 of int64
Ordinary integers (entity ids in value position, tx ids, counts,
journal days, epoch-ms numerics) now use Int64 of int64 instead of
Int of int, eliminating 32-bit int overflow on Melange and removing
the misuse of Instant for plain integers.

- type: Int of int -> Int64 of int64; QueryFormInt of int -> int64
- equality/compare: Int64 vs Float numeric ordering, Int64 vs Instant
  comparable by magnitude but never equal (distinct semantics)
- hashing: murmur3_hash_long for Int64 payloads
- built-ins: exact int64 arithmetic; number?/integer? reject Instant;
  range emits int64 lists
- parser: EDN ints parse via Int64.of_string; #inst -> Instant
- codecs (sqlite + melange): Int64 -> Transit.Int64 unconditionally,
  Instant -> Transit.Date; decode accepts Int/Int64/Big_int and ~m
  tags for backwards compatibility
- storage: normalize_stored_datom migrates legacy Instant values to
  Int64 unless the attr is db.type/instant (real dates preserved)
- js facade: Yojson ints/Intlit -> Int64; Instant still prints as
  #inst/Intlit

+++ bench/bench_ocaml.ml
@@ -184,8 +184,8 @@ let random_man rng i =
         ; "full-name", One_value (String (name ^ " " ^ last_name))
         ; "alias", Many_values alias_values
         ; "sex", One_value (Keyword (if next_int rng 2 = 0 then "male" else "female"))
-        ; "age", One_value (Int (next_int rng 100))
-        ; "salary", One_value (Int (next_int rng 100_000))
+        ; "age", One_value (Int64 (Int64.of_int (next_int rng 100)))
+        ; "salary", One_value (Int64 (Int64.of_int (next_int rng 100_000)))
         ]
     }
 
@@ -255,7 +255,7 @@ let main () =
   bench config "q5-shortcircuit" (fun () ->
     consume_rows
       (q_string
-         ~inputs:[ Arg_scalar (Result_value (String "Anastasia")); Arg_scalar (Result_value (Int 35)) ]
+         ~inputs:[ Arg_scalar (Result_value (String "Anastasia")); Arg_scalar (Result_value (Int64 35L)) ]
          (Lazy.force db)
          "[:find ?e ?n ?l ?a ?s ?al :in $ ?n ?a :where [?e :name ?n] [?e :age ?a] [?e :last-name ?l] [?e :sex ?s] [?e :alias ?al]]"));
   bench config "qpred1" (fun () ->
@@ -263,7 +263,7 @@ let main () =
   bench config "qpred2" (fun () ->
     consume_rows
       (q_string
-         ~inputs:[ Arg_scalar (Result_value (Int 50000)) ]
+         ~inputs:[ Arg_scalar (Result_value (Int64 50000L)) ]
          (Lazy.force db)
          "[:find ?e ?s :in $ ?min-s :where [?e :salary ?s] [(> ?s ?min-s)]]"));
   bench config "q2pred" (fun () ->
